backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.5.3
numpy==1.26.3
pandas==2.1.4
aiosqlite==0.19.0
python-multipart==0.0.6
websockets==12.0
